part3_auto_apply/requirements.txt
playwright>=1.49
python-dotenv
